Add signup page, put CSS in stylesheet, add clicked states for nav buttons, added view cart button to nav
Was struggling with moving my CSS to an external stylesheet for quite a while, when I did, the alignment on the icons in the buttons would break which was very strange. In the end Eujeen helped me to fix it by using !important on the CSS that aligned the icons.